refactor(client): move config loading logic from main.rs to client.rs
This commit centralizes client specific feature and cleans up the main
application entry point.
